requirements.txt
flask==2.3.3
openai==0.28.0
python-dotenv==1.0.0
requests==2.31.0
flask-session==0.5.0
geopy==2.3.0
pytz==2023.3
markupsafe==2.1.3
itsdangerous==2.1.2
werkzeug==2.3.7
jinja2==3.1.2
click==8.1.7